Refactor LoggerService to cast logReasonTypeId to int for improved type consistency in query execution.

+++ grails-app/services/au/org/ala/logger/LoggerService.groovy
@@ -518,7 +518,7 @@ class LoggerService {
                 le("month", toDate)
             }
             if (reasonTypeId) {
-                eq("logReasonTypeId", reasonTypeId)
+                eq("logReasonTypeId", reasonTypeId as int   )
             }
             if (excludeReasonTypeId) {
                 ne("logReasonTypeId", excludeReasonTypeId as int)
